refactor(refactor system): Refactor system monitor widget rendering to improve type consistency ...
# Project State

## Current Focus
Refactor system monitor widget rendering to improve type consistency and reduce potential overflow risks.

## Context
The system monitor widget was refactoring its rendering logic to ensure proper type handling and prevent potential overflow issues during calculations.

## Completed
- [x] Refactored `hist_w` calculation to remove unnecessary `as u16` conversion
- [x] Fixed potential overflow in footer rendering by removing redundant `as u16` cast

## In Progress
- [x] Ongoing work to ensure all rendering calculations maintain proper type safety

## Blockers
- None identified in this change

## Next Steps
1. Verify rendering behavior remains consistent across different terminal sizes
2. Continue refactoring other widget components for similar improvements
---
category: refactor
scope: refactor system
---

+++ examples/_apps/system_monitor.rs
@@ -384,7 +384,7 @@ impl Widget for SystemMonitor {
 
         let cg = self.cpu_gauge.render(Rect::new(0, 0, half_w - 4, gauge_h - 2));
         blit_to(&mut plane, &cg, 2, 3);
-        let hist_w = half_w.saturating_sub(8) as u16;
+        let hist_w = half_w.saturating_sub(8);
         let cpu_color = if self.data.cpu_hist.current() >= 90.0 { t.error }
             else if self.data.cpu_hist.current() >= 70.0 { t.warning } else { t.success };
         render_sparkline(&mut plane, 2, 2 + gauge_h - 2, hist_w.min(20), 2, &self.data.cpu_hist, cpu_color, t.surface);
@@ -450,7 +450,7 @@ impl Widget for SystemMonitor {
         // ── Footer ──
         let footer_y = area.height.saturating_sub(1);
         for x in 0..area.width {
-            let idx = (footer_y * area.width + x as u16) as usize;
+            let idx = (footer_y * area.width + x) as usize;
             if idx < plane.cells.len() { plane.cells[idx].char = '─'; plane.cells[idx].fg = t.outline; }
         }
         let footer = " t:theme  ?:help  ↑↓:nav  q:quit ";
